Remove welcome experience action Make logo switch between light and dark
