+++ crates/project_panel/src/project_panel.rs
@@ -1329,7 +1329,7 @@ impl View for ProjectPanel {
 
                             keystroke_label(
                                 parent_view_id,
-                                "Open project",
+                                "Open a project",
                                 &button_style,
                                 context_menu_item.keystroke,
                                 workspace::Open,
